Make XMLBuilder::DataSet standalone and add getters
This class will be shared with a new XMLReader class that will take care
of turning DOM objects into C++ data structures.
